Move /mem to localhost only
