+++ server/src/http_localhost.go
@@ -34,24 +34,26 @@ func httpLocalhostInit() {
 
 	// Path handlers
 	httpRouter.POST("/ban", httpLocalhostUserAction)
+	httpRouter.POST("/mute", httpLocalhostUserAction)
+	httpRouter.POST("/sendError", httpLocalhostUserAction)
+	httpRouter.POST("/sendWarning", httpLocalhostUserAction)
+	httpRouter.POST("/sendWarningAll", httpLocalhostSendWarningAll)
+
 	httpRouter.GET("/cancel", httpLocalhostCancel)
 	httpRouter.GET("/clearEmptyTables", httpLocalhostClearEmptyTables)
 	httpRouter.GET("/debugFunction", httpLocalhostDebugFunction)
 	httpRouter.GET("/getLongTables", httpLocalhostGetLongTables)
-	httpRouter.GET("/maintenance", httpLocalhostMaintenance)
-	httpRouter.POST("/mute", httpLocalhostUserAction)
-	httpRouter.GET("/print", httpLocalhostPrint)
 	httpRouter.GET("/gracefulRestart", httpLocalhostGracefulRestart)
+	httpRouter.GET("/maintenance", httpLocalhostMaintenance)
+	httpRouter.GET("/mem", httpLocalhostMemory)
+	httpRouter.GET("/print", httpLocalhostPrint)
 	httpRouter.GET("/saveTables", httpLocalhostSaveTables)
-	httpRouter.POST("/sendWarning", httpLocalhostUserAction)
-	httpRouter.POST("/sendWarningAll", httpLocalhostSendWarningAll)
-	httpRouter.POST("/sendError", httpLocalhostUserAction)
 	httpRouter.GET("/shutdown", httpLocalhostShutdown)
 	httpRouter.GET("/terminate", httpLocalhostTerminate)
 	httpRouter.GET("/timeLeft", httpLocalhostTimeLeft)
+	httpRouter.GET("/unmaintenance", httpLocalhostUnmaintenance)
 	httpRouter.GET("/uptime", httpLocalhostUptime)
 	httpRouter.GET("/version", httpLocalhostVersion)
-	httpRouter.GET("/unmaintenance", httpLocalhostUnmaintenance)
 
 	// We need to create a new http.Server because the default one has no timeouts
 	// https://blog.cloudflare.com/the-complete-guide-to-golang-net-http-timeouts/
